Document code with important reminder re React dev mode
In React Dev Mode, the components are mounted twice. That affects
the behavior of the functions that use API.
Per example, there will always be a cooldown since there are
actually 2 API requests that are made to OpenTDB
due to the 2 mounts in React dev mode.
Since we abort the first request (see AbortController in Play.jsx)
we do not see the result of the first request, but the timestamp
of the last request is still persisted. So, the second mount in
dev mode sees a timestamp that is less than 5s away.
To prevent that behavior, either build the project or disable
StrictMode in index.jsx (the latter is not recommended).
In any case, end users will not be affected by these delays that
appear in dev mode.